refactor(cli): split cli.ts into focused modules
Extract 265-line cli.ts into a thin 63-line orchestrator plus three
purpose-built modules under src/cli/:

- auto-restore.ts (49) — snapshot detection + revive prompt
- dispatch.ts (174) — routeComm/routeTools, top aliases, plugin
  registry match, bundled plugin fallback with fuzzy unknown-command
  handling, agent-name shorthand
- error-handler.ts (25) — UserError / AmbiguousMatchError rendering

Pure refactor — no behavior change. import.meta.dir for runBootstrap
stays in cli.ts so the path still resolves to src/.
